edit fake data, edit localstorage

+++ src/localStorage.js
@@ -1,9 +1,13 @@
 
 const saveDonation=title=>{
     const storeDonation = getDonation()
-    storeDonation.push(title)
-    const stringifiedDonation = JSON.stringify(storeDonation)
-    localStorage.setItem("donation", stringifiedDonation)
+    const exists = storeDonation.find(t=>t===title )
+    if(!exists){
+        storeDonation.push(title)
+        const stringifiedDonation = JSON.stringify(storeDonation)
+        localStorage.setItem("donation", stringifiedDonation)
+    }
+
 }
 
 const getDonation=()=>{
